Modularize the rest of the components
